requirements.txt
azure-identity
openai
tiktoken
rouge_score
